requirements.txt
absolufy-imports==0.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.17.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
arxiv==2.2.0
asgiref==3.8.1
asttokens==3.0.0
asyncio==4.0.0
attrs==25.3.0
auth0-python==4.10.0
Authlib==1.6.0
autogen-agentchat==0.7.2
autogen-core==0.7.2
autogen-ext==0.7.2
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
browserbase==1.4.0
build==1.2.2.post1
cachetools==5.5.0
camelot-py==1.0.9
certifi==2026.1.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.1.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.3
crewai==0.201.1
crewai-tools==0.75.0
cryptography==45.0.3
curl_cffi==0.11.1
cycler==0.12.1
databricks-sdk==0.70.0
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
dotenv==0.9.9
duckduckgo_search==8.1.1
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.9
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2024.12.0
generativeai==0.0.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-adk==1.11.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.95.1
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-spanner==3.57.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.30.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.2.4
griffe==1.12.1
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huey==2.5.5
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.8.3
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.1
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
kubernetes==33.1.0
lance-namespace==0.0.6
lance-namespace-urllib3-client==0.0.17
lancedb==0.25.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-deepseek==0.1.4
langchain-google-genai==2.1.9
langchain-huggingface==0.3.1
langchain-ollama==0.3.7
langchain-openai==0.3.30
langchain-perplexity==0.1.2
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.14
litellm==1.74.9
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.2
llama-index-cli==0.5.0
llama-index-core==0.13.2
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.2
llama-index-instrumentation==0.4.0
llama-index-llms-gemini==0.6.0
llama-index-llms-google-genai==0.3.0
llama-index-llms-openai==0.5.4
llama-index-readers-file==0.5.2
llama-index-readers-llama-parse==0.5.0
llama-index-workflows==1.3.0
llama-parse==0.6.54
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.9.2
mdurl==0.1.2
mlflow==3.8.0
mlflow-skinny==3.8.0
mlflow-tracing==3.8.0
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
narwhals==1.45.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.2
oauthlib==3.2.2
ollama==0.5.3
onnxruntime==1.22.0
openai==1.99.9
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
playwright==1.55.0
pluggy==1.6.0
portalocker==2.7.0
posthog==4.10.0
primp==1.1.3
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.38.0
PyMuPDF==1.26.3
pyparsing==3.2.1
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9
safetensors==0.6.2
sarvamai==0.1.25
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==2.3.6
stack-data==0.6.3
stagehand==0.5.3
starlette==0.45.3
streamlit==1.46.1
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uv==0.7.13
uvicorn==0.34.3
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.61
youtube-transcript-api==1.2.2
yt-dlp==2025.2.19
zipp==3.22.0
zstandard==0.24.0
